requirements-web.txt
# Web interface dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
alembic>=1.12.0
redis>=5.0.0
celery>=5.3.0
websockets>=12.0
aiofiles>=23.2.0
jinja2>=3.1.0

# Existing DriftBuddy dependencies
openai>=1.0.0
markdown>=3.4.0
requests>=2.31.0
rich>=13.0.0
structlog>=23.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# LangChain integration
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.1.0
langchain-core>=0.1.0
langchain-text-splitters>=0.1.0

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.25.0 
